mock: split code into modules
