+++ objects/src/mock/block.rs
@@ -1,0 +1,46 @@
+use super::super::{Account, BlockHeader, Digest, Felt, StarkField, Vec, ZERO};
+use crypto::merkle::SimpleSmt;
+use miden_test_utils::rand;
+
+pub fn mock_block_header(
+    block_num: Felt,
+    chain_root: Option<Digest>,
+    note_root: Option<Digest>,
+    accts: &[Account],
+) -> BlockHeader {
+    let acct_db = SimpleSmt::with_leaves(
+        64,
+        accts
+            .iter()
+            .flat_map(|acct| {
+                if acct.is_new() {
+                    None
+                } else {
+                    Some(((*acct.id()).as_int(), *acct.hash()))
+                }
+            })
+            .collect::<Vec<_>>(),
+    )
+    .expect("failed to create account db");
+
+    let prev_hash: Digest = rand::rand_array().into();
+    let chain_root: Digest = chain_root.unwrap_or(rand::rand_array().into());
+    let acct_root: Digest = acct_db.root();
+    let nullifier_root: Digest = rand::rand_array().into();
+    let note_root: Digest = note_root.unwrap_or(rand::rand_array().into());
+    let batch_root: Digest = rand::rand_array().into();
+    let proof_hash: Digest = rand::rand_array().into();
+
+    BlockHeader::new(
+        prev_hash,
+        block_num,
+        chain_root,
+        acct_root,
+        nullifier_root,
+        note_root,
+        batch_root,
+        proof_hash,
+        ZERO,
+        rand::rand_value(),
+    )
+}
